% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_analysis.R
\name{clean.default}
\alias{clean.default}
\title{Default cleanup for eDNA water}
\usage{
clean.default(
  ps_obj,
  taxa_minR = 0,
  taxa_minP = 0,
  taxa_select = NA,
  taxa_excl = "Homo"
)
}
\arguments{
\item{ps_obj}{Phyloseq object}

\item{taxa_minR}{Minimun number of reads for a taxon to be retained (default=0)}

\item{taxa_minP}{Minimun RRA for a taxon to be retained (default=0)}

\item{taxa_select}{Select specific taxa, can be of any level. Can select multiple taxa with "taxA|taxB"}

\item{taxa_excl}{Exclude certain taxa, can be of any level. Can exclude multiple taxa with "taxA|taxB". (default="Homo")}
}
\value{
Phyloseq object
}
\description{
Clean data for analysis: select certain taxa, remove certain taxa, threshold the data and transform reads to RRA
}
\examples{
#To add
}
