% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_analysis.R
\name{export.data_sets}
\alias{export.data_sets}
\title{Export eDNA water datasets}
\usage{
export.data_sets(
  sample_sheet_df,
  otu_df,
  tax_df,
  ID_cutoff = 1,
  out_path = "."
)
}
\arguments{
\item{sample_sheet_df}{Data frame of (lab) sample sheet}

\item{otu_df}{Data frame of OTU/ASV table}

\item{tax_df}{Data frame with taxonomic annotation data}

\item{ID_cutoff}{Reference database similarity needed to filter OTU/ASV data.}

\item{out_path}{Directory to store the plots in. (Default = ".")}
}
\value{
List of ggplot2 objects if out_path is not defined.
}
\description{
Convert "raw" operational data to input for data analysis
}
\examples{
#To add
}
