% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{extract.fastqfilter_log.readcount}
\alias{extract.fastqfilter_log.readcount}
\title{Read read-counts from logfile - VSEARCH - fastq_filter}
\usage{
extract.fastqfilter_log.readcount(logfile)
}
\arguments{
\item{logfile}{Path to a VSEARCH log file}
}
\value{
Integer with total readcount of sample **after** VSEARCH - fastq_filter (i.e. qc-filtered)
}
\description{
(version 2.22.1)
}
\examples{
#To add
}
