% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{extract.mergepairs_log.tot_input}
\alias{extract.mergepairs_log.tot_input}
\title{Read vsearch --fastq_mergepairs log file to get total input reads}
\usage{
extract.mergepairs_log.tot_input(logfile)
}
\arguments{
\item{logfile}{Path to a VSEARCH log file}
}
\value{
Integer with total readcount of sample **before** VSEARCH --mergepairs (i.e. paired-end input)
}
\description{
(version 2.22.1)
}
\examples{
#To add
}
