% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{get.sample_sheet_urls}
\alias{get.sample_sheet_urls}
\title{Get sample sheets URLs from LST002}
\usage{
get.sample_sheet_urls()
}
\value{
List with URLs
}
\description{
Read sample sheet URLs from \href{https://docs.google.com/spreadsheets/d/1h70fKsQCqPQS5tSQYOwws2zeCJJu9BOlH7mI0xUg5rA/edit?gid=1637141996#gid=1637141996}{LST002}
}
\examples{
# todo
}
