% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_analysis.R
\name{make.default_project_barplots}
\alias{make.default_project_barplots}
\title{Create default barplots for project level QC}
\usage{
make.default_project_barplots(ps_obj, RRA = F, out_path = NA)
}
\arguments{
\item{ps_obj}{Phyloseq object as INPUT}

\item{RRA}{Convert reads to relative read abundance (RRA) per sample. (Default FALSE)}

\item{out_path}{[RECOMMENDED] Directory to store the plots in. For projects with a lot of samples this should always be used. Otherwise your R-session might crash!}
}
\value{
List of ggplot2 objects if out_path is not defined.
}
\description{
Pass run level data to generate project level barplots with some pre-configured parameters to visualize an overview, a fish and amphibian specific plot and a most abundant species plot.
}
\examples{
#To add
}
