% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{merger.taxonomy_usearch}
\alias{merger.taxonomy_usearch}
\title{Merge VSEARCH BLAST6 and LCA format}
\usage{
merger.taxonomy_usearch(usearch_blast6, usearch_lca)
}
\arguments{
\item{usearch_blast6}{Path/URL to, or data.frame with, the BLAST6 output}

\item{usearch_lca}{Path/URL to, or data.frame with, the LCA output}
}
\value{
Dataframe with aggregated BLAST6 taxonomy information (columns) of target OTU
}
\description{
Merging BLAST6 (multiple RefDB hits per query/OTU) with LCA output (1 taxonomy per query/OTU) to generate OBITools like taxonomy table.

BLAST6 data of multiple hits are collapsed and represented as a single element per query/OTU row.
}
\examples{
# to add
}
