% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{parse.utax_string}
\alias{parse.utax_string}
\title{Parse VSEARCH SINTAX taxonomy string to dataframe}
\usage{
parse.utax_string(x, is_pred = F)
}
\arguments{
\item{x}{VSEARCH SINTAX taxonomy string}

\item{is_pred}{(SINTAX output specific) If bootstrap values are part of the tax_string. Default = FALSE.}
}
\value{
Data.frame with (1) row and (n = tax_levels) columns
}
\description{
Each taxonomic identifier must start with an indication of the rank by one of the letters d (for domain) k (kingdom), p (phylum), c (class), o (order), f (family), g (genus), s (species), or t (strain). The letter is followed by a colon (:) and the name of that rank. Commas and semicolons are not allowed in the name of the rank. Non-ascii characters should be avoided in the names.

Example:

>X80725_S000004313;tax=d:Bacteria,p:Proteobacteria,c:Gammaproteobacteria,o:Enterobacteriales,f:Enterobacteriaceae,g:Escherichia/Shigella,s:Escherichia_coli,t:str._K-12_substr._MG1655
}
\examples{
tax_string = "d:Bacteria,p:Proteobacteria,c:Gammaproteobacteria,o:Enterobacteriales,f:Enterobacteriaceae,g:Escherichia/Shigella,s:Escherichia_coli,t:str._K-12_substr._MG1655"
tax_string |> parse.utax_string()
}
