% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_analysis.R
\name{ps.clean_basic}
\alias{ps.clean_basic}
\title{Basic cleanup for eDNA water samples}
\usage{
ps.clean_basic(
  ps_obj,
  RRA = T,
  taxa_minRS = 0,
  taxa_minPS = 0,
  taxa_minR = 0,
  taxa_minP = 0,
  taxa_select = NA,
  taxa_excl = "Homo",
  sample_minR = 0
)
}
\arguments{
\item{ps_obj}{Phyloseq object}

\item{RRA}{Divide taxa values by sample total. In context of reads this results in Relative Read Abundances (RRA)}

\item{taxa_minRS}{Minimum number of reads for a taxon to be retained in a sample (default=0)}

\item{taxa_minPS}{Minimum RRA for a taxon to be retained in a sample (default=0)}

\item{taxa_minR}{Minimum number of reads for a taxon to be retained in the complete dataset (default=0). Note that 0 is a special case, as all data with 0 reads always gets removed, so this translates to > 0. For other values it translates as >= x. This also applies to other thresholding arguments.}

\item{taxa_minP}{Minimum RRA for a taxon to be retained the complete dataset (default=0)}

\item{taxa_select}{Select specific taxa, can be of any level. Can select multiple taxa with "taxA|taxB"}

\item{taxa_excl}{Exclude certain taxa, can be of any level. Can exclude multiple taxa with "taxA|taxB". (default="Homo")}

\item{sample_minR}{Minimum number of reads for a sample to be retained (default=0)}
}
\value{
Phyloseq object
}
\description{
Clean data for analysis: select and exclude taxa, threshold the data and transform reads to RRA. Global thresholds are applied before sample-specific thresholds. Taxa selection is applied before thresholding. RRA is performed as the final step. Thresholds are applied at the same time, with the most restrictive prevailing. Completely empty samples and/or taxa are always removed.
}
\examples{
#To add
}
