% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_analysis.R
\name{ps.get_otu_tax}
\alias{ps.get_otu_tax}
\title{Extract OTU table and taxonomy table from phyloseq object}
\usage{
ps.get_otu_tax(ps)
}
\arguments{
\item{ps}{Phyloseq object}
}
\value{
Data frame
}
\description{
Extract OTU and taxonomy data from phyloseq object.
}
\examples{
#To add
}
