% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot.basic_exploration}
\alias{qcplot.basic_exploration}
\title{Basic exploration plots}
\usage{
qcplot.basic_exploration(ps_obj)
}
\arguments{
\item{ps_obj}{Phyloseq object as INPUT}
}
\value{
List with ggplot objects
}
\description{
Plots of basic statistics of th operational data
}
\examples{
#To add
}
