% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot.plate_column_asv_analysis}
\alias{qcplot.plate_column_asv_analysis}
\title{[WRAPPER] Barplot and ordination for QC on possible strip-swaps}
\usage{
qcplot.plate_column_asv_analysis(ps_obj, topX = 25, out_path = NA)
}
\arguments{
\item{ps_obj}{Phyloseq object as INPUT}

\item{topX}{integer, topX taxa per COLUMN on the plate (default=25)}

\item{out_path}{Directory to store the plots in. (If this argument is used the default list will not be returned.)}
}
\value{
List with n (plates) objects, for each a barplot and ordination
}
\description{
Taxon-barplot and ordination with data grouped per plate and per column to investigate lab-specific errors/trends
}
\examples{
#To add
}
