% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot.plate_heatmap_toptaxa}
\alias{qcplot.plate_heatmap_toptaxa}
\title{Top Taxonomy Density Heatmap for QC}
\usage{
qcplot.plate_heatmap_toptaxa(ps_obj, omit_cutoff = 100)
}
\arguments{
\item{ps_obj}{Phyloseq object as INPUT}

\item{omit_cutoff}{integer, omit samples with less than n reads (default=100). This is important for color gradient to make sense.}
}
\value{
ggplot2::geom_tile()
}
\description{
Map most abundant ASV per sample to plate layout to investigate lab-specific errors/trends
}
\examples{
#To add
}
