% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot.sample_barplot_taxonomic_coverage}
\alias{qcplot.sample_barplot_taxonomic_coverage}
\title{Check taxonomic assignment levels per sample}
\usage{
qcplot.sample_barplot_taxonomic_coverage(ps_obj)
}
\arguments{
\item{ps_obj}{Phyloseq object as INPUT}
}
\value{
ggplot2 object
}
\description{
Per sample the read % attributed to each taxonomic level
}
\examples{
#To add
}
