% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read.fastx_gdrive}
\alias{read.fastx_gdrive}
\title{Try to read a FASTX file from File Path or google Drive}
\usage{
read.fastx_gdrive(file_id, tempf = tempfile(), ...)
}
\arguments{
\item{file_id}{Path or URL (Google Drive) to a FASTX file that can be read with \link[Biostrings]{readDNAStringSet}}

\item{tempf}{File path for downloaded data (default=\link[base]{tempfile})}

\item{...}{See \link[Biostrings]{readDNAStringSet}}
}
\value{
\link[Biostrings]{DNAStringSet}
}
\description{
At INBO Google Drive is the main platform for RDM. This function reads FASTX files into \link[Biostrings]{DNAStringSet} from Google Drive link
(right-click file, "Share", "Copy link")
}
\examples{
#To add
}
