% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{read.idtaxa_taxonomy}
\alias{read.idtaxa_taxonomy}
\title{Read DECIPHER IdTaxa() taxonomy output into dataframe}
\usage{
read.idtaxa_taxonomy(
  tax_tab_path,
  read_supported_only = TRUE,
  add_source = FALSE,
  clean_tax = FALSE
)
}
\arguments{
\item{tax_tab_path}{Path to taxonomy CSV table file}

\item{read_supported_only}{Only positive annotations (default=TRUE)}

\item{add_source}{Add basename of `tax_tab_path` to output dataframe (default=FALSE)}

\item{clean_tax}{Try to remove non-informative text in taxonomic annotations (default=FALSE)}
}
\value{
Dataframe with taxonomy information (columns) per OTU (rows)
}
\description{
Reading and general cleaning/pruning of DECIPHER IdTaxa() output
}
\examples{
#To add
}
