% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{read.log_mumu}
\alias{read.log_mumu}
\title{Read MUMU log and return in dataframe}
\usage{
read.log_mumu(mumu_log_path)
}
\arguments{
\item{mumu_log_path}{Path to MUMU logfile (can also be link to Googe Drive)}
}
\value{
Dataframe with MUMU log
}
\description{
Read mumu log, add column names based on documentation.
https://github.com/frederic-mahe/mumu/blob/09a510bca992562405f1904debfbd10ae8d7bdef/man/mumu.1#L124
}
\details{
> Output file for OTU merging statistics (18 columns separated by tabulations, first line is a header line with column names)

- This is incorrect, no header is present in the file. -> Hence this function.
}
\examples{
#To add
}
