% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{read.logs_preprocess}
\alias{read.logs_preprocess}
\title{Read multiple log files for pre-processing workflow}
\usage{
read.logs_preprocess(mydir, stepname, logpattern = ".log")
}
\arguments{
\item{mydir}{Path to a folder containing .log files from pre-processing}

\item{stepname}{One of: "merged", "trimmed" or "filter"}

\item{logpattern}{Pattern to find correct log files, see \link[base]{list.files} (default = ".log")}
}
\value{
Dataframe with readcounts per file
}
\description{
Wrapper to extract the remaining readcounts from .log files for each step of the INBO eDNA metabarcoding pre-processing workflow
Implemented for:
  - merged (VSEARCH --fastq_mergepairs)
  - primer-trimmed (cutadapt)
  - QC filtered (VSEARCH --fastq_filter)
}
\examples{
#To add
}
