% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read.LST002}
\alias{read.LST002}
\title{Read sequencing-run overview list (LST002) from Google Drive}
\usage{
read.LST002()
}
\value{
Data.frame with LST002 content, all data converted to type character
}
\description{
Team Genetic Diversity (INBO) uses the Google Drive as main platform for RDM.
This function reads the sequencing-run overview (LST002) Google sheet.
}
\details{
Can only access this data with a working INBO Google-account, that has sufficient permissions/access.
}
\examples{
# read.LST002()
}
\references{
LST002: \url{https://docs.google.com/spreadsheets/d/1h70fKsQCqPQS5tSQYOwws2zeCJJu9BOlH7mI0xUg5rA}
}
