% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{read.otu_table_vsearch}
\alias{read.otu_table_vsearch}
\title{Read VSEARCH OTU-tab output into dataframe}
\usage{
read.otu_table_vsearch(otu_table_path)
}
\arguments{
\item{otu_table_path}{Path to OTU table file (can also be link to Googe Drive)}
}
\value{
Dataframe with read counts per (z)OTU (rows) per sample (columns)
}
\description{
Read OTU table, cleanup colnames and return
}
\examples{
#To add
}
