% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{read.sample_sheet_lab}
\alias{read.sample_sheet_lab}
\title{Read lab-sample-sheet}
\usage{
read.sample_sheet_lab(sample_metadata_path)
}
\arguments{
\item{sample_metadata_path}{Filepath for \link[readxl]{read_xlsx} or google-sheets id \link[googlesheets4]{range_read}}
}
\value{
Very basic cleaned and formatted dataframe from lab-sample-sheet
}
\description{
All purpose starting point for reading the lab sample_sheet.
}
\examples{
# example on public available data
# googlesheets4::gs4_deauth()
# my_sample_data_df = read.sample_sheet_lab("1cWkVqk3y7668OVRhIlpa0fUmTh4oJTRWUqp60GRJ5J8")
}
