% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{read.taxonomy_boldigger3}
\alias{read.taxonomy_boldigger3}
\title{Read BOLDigger3 taxonomy output into dataframe}
\usage{
read.taxonomy_boldigger3(tax_tab_path, add_source = FALSE, clean_tax = FALSE)
}
\arguments{
\item{tax_tab_path}{Path to taxonomy XLSX table file}

\item{add_source}{Add basename of `tax_tab_path` to output dataframe (default=FALSE)}

\item{clean_tax}{Try to remove non-informative text in taxonomic annotations (default=FALSE)}
}
\value{
Dataframe with taxonomy information (columns) per OTU (rows)
}
\description{
Read BOLDigger3 identification_result.xlsx output
}
\examples{
#To add
}
