% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{read.taxonomy_sintax}
\alias{read.taxonomy_sintax}
\title{Read VSEARCH SINTAX taxonomy output into dataframe}
\usage{
read.taxonomy_sintax(
  sintax_path,
  read_supported_only = TRUE,
  add_source = FALSE,
  clean_tax = FALSE,
  add_custom = F
)
}
\arguments{
\item{sintax_path}{Path to taxonomy table file}

\item{read_supported_only}{Only return bootstrap supported annotations (default=TRUE)}

\item{add_source}{Add basename of `sintax_path` to output dataframe (default=FALSE)}

\item{clean_tax}{Try to remove non-informative text in taxonomic annotations (default=FALSE)}

\item{add_custom}{Add a custom taxonomy-level between Phylum and Class with common names for basic quality control (INBO specific) (default=FALSE)}
}
\value{
Dataframe with taxonomy information (columns) per OTU (rows)
}
\description{
Reading and general cleaning/pruning of VSEARCH SINTAX output
}
\examples{
#To add
}
