% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{read.taxonomy_usearch_blast6}
\alias{read.taxonomy_usearch_blast6}
\title{Read VSEARCH BLAST6 format to dataframe}
\usage{
read.taxonomy_usearch_blast6(ublast6_path)
}
\arguments{
\item{ublast6_path}{Path to ublast6 output file (tsv). Can be file_path or google-drive URL}
}
\value{
Dataframe with taxonomy information (columns) of target per query hit
}
\description{
Reading and basic formatting of vsearch --blast6 output
}
\examples{
# to add
}
