% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{read.taxonomy_usearch_lca}
\alias{read.taxonomy_usearch_lca}
\title{Read VSEARCH LCA format to dataframe}
\usage{
read.taxonomy_usearch_lca(lca_path)
}
\arguments{
\item{lca_path}{Path to lca output file (tsv). Can be file_path or google-drive URL}
}
\value{
Dataframe with taxonomy information (columns) of target per query hit
}
\description{
Reading and basic formatting of vsearch --lca_out output
}
\examples{
# to add
}
