% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{read.vsearch_otu_table}
\alias{read.vsearch_otu_table}
\title{Read VSEARCH OTU-tab output into dataframe}
\usage{
read.vsearch_otu_table(otu_table_path)
}
\arguments{
\item{otu_table_path}{Path to OTU table file}
}
\value{
Dataframe with read counts per (z)OTU (rows) per sample (columns)
}
\description{
Read OTU table, cleanup colnames and return
}
\examples{
#To add
}
