% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helper.R
\name{report.demultiplex}
\alias{report.demultiplex}
\title{Render Rmarkdown Report - Demultiplexing}
\usage{
report.demultiplex(
  sample_sheet_url,
  demultiplex_result_path,
  out_dir = "demultiplex-QC-report"
)
}
\arguments{
\item{sample_sheet_url}{Filepath for \link[readxl]{read_xlsx} or google-sheets id \link[googlesheets4]{range_read}}

\item{demultiplex_result_path}{HPC-ONLY: file_path to demultiplexing results}

\item{out_dir}{Output folder (Default="demultiplex-QC-report")}
}
\value{
HTML report for demultiplexing results, in `out_dir`
}
\description{
//////// HPC-ONLY //////////
}
\examples{
#To add
}
