% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helper.R
\name{report.nextits1}
\alias{report.nextits1}
\title{Render Rmarkdown Report - NextITS Step1}
\usage{
report.nextits1(
  sample_sheet_url,
  nextits1_result_path,
  out_dir = "NextITS-Step1-QC-report"
)
}
\arguments{
\item{sample_sheet_url}{Filepath for \link[readxl]{read_xlsx} or google-sheets id \link[googlesheets4]{range_read}}

\item{nextits1_result_path}{HPC-ONLY: file_path to NextITS step1 results}

\item{out_dir}{Output folder (Default="pre-process-QC-report")}
}
\value{
HTML report for NextITS Step1 results, in `out_dir`
}
\description{
//////// HPC-ONLY //////////
}
\examples{
#To add
}
