% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helper.R
\name{report.samplesheet}
\alias{report.samplesheet}
\title{Render Rmarkdown Report - Lab Sample Sheet}
\usage{
report.samplesheet(sample_sheet_url, out_dir = "sample-sheet-QC-report")
}
\arguments{
\item{sample_sheet_url}{Filepath for \link[readxl]{read_xlsx} or google-sheets id \link[googlesheets4]{range_read}}

\item{out_dir}{Output folder (Default="sample-sheet-QC-report")}
}
\value{
HTML report and barcode-files for demultiplexing, in `out_dir`
}
\description{
//////////////////
}
\examples{
#To add
}
