% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{select.obitools3_asv_taxonomy}
\alias{select.obitools3_asv_taxonomy}
\title{Filter ASV data for eDNA-water (12S primers)}
\usage{
select.obitools3_asv_taxonomy(tax_df, min_BEST_ID = 1, max_BEST_ID = 1)
}
\arguments{
\item{tax_df}{Dataframe of (obitools3) taxonomy table (see \link[gendiver]{read.obitools3_taxonomy})}

\item{min_BEST_ID}{Minimum identity match for ASV to be retained. Default=1, retain only ASVs with perfect match to RefDB}

\item{max_BEST_ID}{Maximum identity match for ASV to be retained. Default=1, retain only ASVs with perfect match to RefDB}
}
\value{
Dataframe with ASV selection
}
\description{
Basic pruning of OBITools3 taxonomy dataframe (for RIAZ primer), formatting, adding custom taxon for team gendiv,
adding unclassified labels, cleaning and converting to phyloseq tax_table() compatible dataframe
}
\examples{
#To add
}
