% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_analysis.R
\name{data.merge_replicates}
\alias{data.merge_replicates}
\title{Merge technical replicates into samples}
\usage{
data.merge_replicates(
  otu_df,
  tax_df,
  sample_sheet_df,
  merge_col = "INDIVIDUAL",
  min_rprev = 1
)
}
\arguments{
\item{otu_df}{Data frame of OTU/ASV table}

\item{tax_df}{Data frame with taxonomic annotation data}

\item{sample_sheet_df}{Data frame of (lab) sample sheet}

\item{merge_col}{Column name of `sample_sheet_df` to merge on. (default="INDIVIDUAL")}

\item{min_rprev}{Optional filter to exclude taxa that occur in less than n replicates. Default=1 (keep all taxa)}
}
\value{
phyloseq object
}
\description{
Merge the data from (generally 3) replicate amplicons into a single sample. Here, this is done by taking the average of each replicate's relative read abundance (RRA).
}
\examples{
#To add
}
