% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{extract.cutadapt_log.readcount}
\alias{extract.cutadapt_log.readcount}
\title{Read read-counts from logfile - CUTADAPT}
\usage{
extract.cutadapt_log.readcount(logfile)
}
\arguments{
\item{logfile}{Path to a CUTADAPT log file}
}
\value{
Integer with readcount of sample **after** CUTADAPT (i.e. trimmed)
}
\description{
cutadapt --version 5.0
}
\examples{
#To add
}
