% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{extract.mergepairs_log.readcount}
\alias{extract.mergepairs_log.readcount}
\title{Read read-counts from logfile - VSEARCH --mergepairs}
\usage{
extract.mergepairs_log.readcount(logfile)
}
\arguments{
\item{logfile}{Path to a VSEARCH log file}
}
\value{
Integer with readcount of sample **after** VSEARCH --mergepairs (i.e. merged)
}
\description{
(version 2.22.1)
}
\examples{
#To add
}
