% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{extract.pear_log.readcount}
\alias{extract.pear_log.readcount}
\title{Read read-counts from logfile - PEAR merging}
\usage{
extract.pear_log.readcount(logfile)
}
\arguments{
\item{logfile}{Path to a PEAR log file}
}
\value{
Integer with readcount of sample **after** PEAR merging (i.e. merged)
}
\description{
Read read-counts from logfile - PEAR merging
}
\examples{
#To add
}
