% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_analysis.R
\name{make.taxa_barplot}
\alias{make.taxa_barplot}
\title{General barplot function for taxa inspection}
\usage{
make.taxa_barplot(
  ps_obj,
  taxa_rank = "species",
  min_percent_total = 0,
  RRA = FALSE
)
}
\arguments{
\item{ps_obj}{Phyloseq object as INPUT}

\item{taxa_rank}{Taxonomic rank to visualise the data on}

\item{min_percent_total}{numeric, remove taxa that have less than n percent reads in the dataset.}

\item{RRA}{Convert reads to relative read abundance (RRA) per sample. (Default FALSE)}
}
\value{
ggplot2 object
}
\description{
Quickly visualize taxa groups in a barplot
}
\examples{
#To add
}
