% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_analysis.R
\name{make.taxa_barplot}
\alias{make.taxa_barplot}
\title{General barplot function for taxa inspection}
\usage{
make.taxa_barplot(
  ps_obj,
  taxa_rank = "species",
  taxa_select = NA,
  taxa_excl = "Homo",
  cutoff = 1,
  RRA = FALSE
)
}
\arguments{
\item{ps_obj}{Phyloseq object as INPUT}

\item{taxa_rank}{Taxonomic rank to visualise the data on}

\item{taxa_select}{String to select the data to plot based on taxonomy. Multiple taxonomies can be passed using "|". Taxonomies do not have to be of the same level, and are case insensitive. E.g. taxa_select='Amphibia|Homo sapiens`}

\item{taxa_excl}{String to select the taxonomies to exclude from the data before plotting. Same logic applies as for taxa_select. (Default 'Homo')}

\item{cutoff}{numeric, remove taxa that have less than n percent reads in the dataset.}

\item{RRA}{Convert reads to relative read abundance (RRA) per sample. (Default FALSE)}
}
\value{
ggplot2 object
}
\description{
Quickly visualize taxa groups in a barplot
}
\examples{
#To add
}
