% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{map.filetype2tool}
\alias{map.filetype2tool}
\title{Map pre-processing output files to used tool}
\usage{
map.filetype2tool(x)
}
\arguments{
\item{x}{Filename (string)}
}
\value{
One of: "saber-demultiplexing", "pear-merging", "vsearch-merging", "cutadapt-trimming" or "vsearch-quality-filter"
}
\description{
In the INBO eDNA emtabarcoding workflow files have structured filenames. Translate these filenames to the used tool.
}
\examples{
#To add
}
