% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{propagate.unclassified_taxonomy}
\alias{propagate.unclassified_taxonomy}
\title{Complete NA identifications in taxonomic annotation results}
\usage{
propagate.unclassified_taxonomy(tax, uncl_suffix = "_unclassified")
}
\arguments{
\item{tax}{Taxonomy table}

\item{uncl_suffix}{Suffix to use in propagation (default="_unclassified")}
}
\value{
Dataframe with taxonomy information (columns) per OTU (rows)
}
\description{
Complete (OBITools3) taxonomy by replacing NA values with last known taxon + suffix
}
\examples{
#To add
}
