% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_analysis.R
\name{ps.merge_replicates}
\alias{ps.merge_replicates}
\title{Merge technical replicates into samples}
\usage{
ps.merge_replicates(ps_obj)
}
\arguments{
\item{ps_obj}{phyloseq object with sample_data() containing "INDIVIDUAL"}
}
\value{
phyloseq object
}
\description{
Merge the data from (generally 3) replicate amplicons into a single sample. Here, this is done by taking the average of each replicate's relative read abundance (RRA).
}
\examples{
#To add
}
