% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_analysis.R
\name{ps.merge_replicates}
\alias{ps.merge_replicates}
\title{Merge technical replicates into samples}
\usage{
ps.merge_replicates(otu_df, tax_df, sample_sheet_df)
}
\arguments{
\item{otu_df}{Data frame of OTU/ASV table}

\item{tax_df}{Data frame with taxonomic annotation data}

\item{sample_sheet_df}{Data frame of (lab) sample sheet}
}
\value{
phyloseq object
}
\description{
Merge the data from (generally 3) replicate amplicons into a single sample. Here, this is done by taking the average of each replicate's relative read abundance (RRA).
}
\examples{
#To add
}
