% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{pull.taxonomy_obitools3}
\alias{pull.taxonomy_obitools3}
\title{Read and Filter ASV data for eDNA-water (12S primers)}
\usage{
pull.taxonomy_obitools3(tax_table_path, min_BEST_ID = 1, max_BEST_ID = 1)
}
\arguments{
\item{tax_table_path}{Path to taxonomy table file (can also be link to Googe Drive)}

\item{min_BEST_ID}{Minimum identity match for ASV to be retained. Default=1, retain only ASVs with perfect match to RefDB}

\item{max_BEST_ID}{Maximum identity match for ASV to be retained. Default=1, retain only ASVs with perfect match to RefDB}
}
\value{
Dataframe with ASV selection
}
\description{
Wrapper for reading and cleaning obitools3 tax table to phyloseq tax_table() compatible dataframe
}
\examples{
#To add
}
