% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot.plate_column_asv_barplot}
\alias{qcplot.plate_column_asv_barplot}
\title{Barplot and ordination for QC on possible strip-swaps}
\usage{
qcplot.plate_column_asv_barplot(ps_plate, topX = 25)
}
\arguments{
\item{ps_plate}{Phyloseq object as INPUT  (Ideally of 1 plate)}

\item{topX}{integer, topX taxa per COLUMN on the plate (default=25)}
}
\value{
List with 2 plots: barplot and ordination
}
\description{
Taxon-barplot and ordination with data grouped per plate and per column to investigate lab-specific errors/trends
}
\examples{
#To add
}
