% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot.plate_heatmap_readcount}
\alias{qcplot.plate_heatmap_readcount}
\title{Total Readcount Density Heatmap for QC}
\usage{
qcplot.plate_heatmap_readcount(
  ps_obj,
  cutoff = 0,
  transform = "identity",
  mark_controls = FALSE,
  mark_data = FALSE,
  all_wells = FALSE
)
}
\arguments{
\item{ps_obj}{Phyloseq object as INPUT}

\item{cutoff}{numeric, omit samples with a value of n or lower (default=0, so more than 0).}

\item{transform}{(default="identity") See \link[ggplot2]{scale_continuous}, e.g. "log2", "log10", "sqrt", ...}

\item{mark_controls}{(default=FALSE) Mark tiles based on ANALYSE_TYPE. Only non MB data.}

\item{mark_data}{(default=FALSE) Mark tiles based on ANALYSE_TYPE. All experiment/MB data.}

\item{all_wells}{(default=FALSE) Plot all wells, regardless if data is present or not. If turned off (`mark_all=FALSE`, default), only plot wells with data above the omit cutoff.}
}
\value{
ggplot2::geom_tile()
}
\description{
Map readcounts per sample to plate layout to investigate lab-specific errors/trends
Recommended to first make a appropriate phyloseq object of the data you want to investigate
Example: PAC bleed-over analysis -> input phyloseq is of all exotic ASVs over all samples.
}
\examples{
#To add
}
