% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot.plate_heatmap_readcount}
\alias{qcplot.plate_heatmap_readcount}
\title{Total Readcount Density Heatmap for QC}
\usage{
qcplot.plate_heatmap_readcount(
  ps_obj,
  omit_cutoff = 100,
  transform = "identity"
)
}
\arguments{
\item{ps_obj}{Phyloseq object as INPUT}

\item{omit_cutoff}{integer, omit samples with less than n reads (default=100). This is important for color gradient to make sense.}

\item{transform}{(default="identity") See \link[ggplot2]{scale_continuous}, e.g. "log2", "log10", "sqrt", ...}
}
\value{
ggplot2::geom_tile()
}
\description{
Map readcounts per sample to plate layout to investigate lab-specific errors/trends
Recommended to first make a appropriate phyloseq object of the data you want to investigate
Example: PAC bleed-over analysis -> input phyloseq is of all exotic ASVs over all samples.
}
\examples{
#To add
}
