% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex.R
\name{read.summary_files_demultiplex}
\alias{read.summary_files_demultiplex}
\title{Read per-sample readcounts from summary files after demultiplexing job}
\usage{
read.summary_files_demultiplex(data_dir, summary_file_pattern = "sample.txt")
}
\arguments{
\item{data_dir}{File path to the `02_demultiplex/` dir}

\item{summary_file_pattern}{Pattern to recognize summary files (default="sample.txt)}
}
\value{
Dataframe with read counts per sample
}
\description{
Find and read auto-generated demultiplexing results from demultiplexing folder
}
\examples{
#To add
}
