% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read.table_gdrive}
\alias{read.table_gdrive}
\title{Try to read a table from File Path or google Drive}
\usage{
read.table_gdrive(file_id, ...)
}
\arguments{
\item{file_id}{Path or URL (Google Drive) to a tabular file that can be read with \link[utils]{read.table}}

\item{...}{See \link[utils]{read.table}}
}
\value{
Dataframe with MUMU log
}
\description{
At INBO Google Drive is the main platform for RDM. This functions can read CSV tables from Google Drive link
(right-click file, "Share", "Copy link")
}
\examples{
#To add
}
