% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{read.taxonomy_obitools3}
\alias{read.taxonomy_obitools3}
\title{Read OBITools3 ECOTAG taxonomy output into dataframe}
\usage{
read.taxonomy_obitools3(tax_table_path)
}
\arguments{
\item{tax_table_path}{Path to taxonomy table file (can also be link to Googe Drive)}
}
\value{
Dataframe with taxonomy information (columns) per ASV (rows)
}
\description{
Reading and general cleaning/pruning of OBITOOLS3 ecotag output
}
\examples{
#To add
}
