% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helper.R
\name{report.operational_data}
\alias{report.operational_data}
\title{Render Rmarkdown Report - Basic-analysis Clustering and Taxonomy}
\usage{
report.operational_data(
  sample_sheet_url,
  tax_file_path,
  otu_table_path,
  seq_fasta_path,
  out_dir = "operational-data-QC-report"
)
}
\arguments{
\item{sample_sheet_url}{Filepath for \link[readxl]{read_xlsx} or google-sheets id \link[googlesheets4]{range_read}}

\item{tax_file_path}{Path to taxonomy table file (can also be link to Googe Drive)}

\item{otu_table_path}{Path to OTU table file (can also be link to Googe Drive)}

\item{seq_fasta_path}{Path or URL (Google Drive) to a FASTX file that can be read with \link[Biostrings]{readDNAStringSet}}

\item{out_dir}{Output folder (Default="operational-data-QC-report")}
}
\value{
HTML report for Clustering and Taxonomy results, in `out_dir`
}
\description{
//////// //////////
}
\examples{
#To add
}
