% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helper.R
\name{report.preprocess}
\alias{report.preprocess}
\title{Render Rmarkdown Report - Pre-Processing}
\usage{
report.preprocess(
  sample_sheet_url,
  preprocess_result_path,
  out_dir = "pre-process-QC-report"
)
}
\arguments{
\item{sample_sheet_url}{Filepath for \link[readxl]{read_xlsx} or google-sheets id \link[googlesheets4]{range_read}}

\item{preprocess_result_path}{HPC-ONLY: file_path to pre-processing results}

\item{out_dir}{Output folder (Default="pre-process-QC-report")}
}
\value{
HTML report for demultiplexing results, in `out_dir`
}
\description{
//////// HPC-ONLY //////////
}
\examples{
#To add
}
