% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex.R
\name{summarize.readcounts_demultiplex}
\alias{summarize.readcounts_demultiplex}
\title{Summarize demultiplexing readcounts for plotting}
\usage{
summarize.readcounts_demultiplex(
  df,
  formula = read_count ~ type + multiplexed_source
)
}
\arguments{
\item{df}{Data frame with read counts per sample (output of \link[gendiver]{read.summary_files_demultiplex})}

\item{formula}{Formula to use for aggregation (default: `read_count ~ type + multiplexed_source`)}
}
\value{
Summary data ready for plotting
}
\description{
demultiplexed readcounts per sample
}
\examples{
#To add
}
