% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{write.barcode_files}
\alias{write.barcode_files}
\title{Create barcode-file for demultiplexing}
\usage{
write.barcode_files(lab_sample_sheet, out_dir, LIB_COL = 1)
}
\arguments{
\item{lab_sample_sheet}{Dataframe with lab-sample-data (see \link[gendiver]{read.lab_sample_sheet})}

\item{out_dir}{Directory path to write barcode-files to}

\item{LIB_COL}{Column name or ID designating the multiplexed POOLs}
}
\value{
Nothing, prints POOL-ids and writes files to `outdir`
}
\description{
Loop over all multiplexed sample names (POOLS) and write TSV with USI - barcodes combinations
Check the lab-sample-sheet you use that each USI-barcode combination is UNIQUE!
}
\examples{
my_sample_data_df = read.lab_sample_sheet("1cWkVqk3y7668OVRhIlpa0fUmTh4oJTRWUqp60GRJ5J8")
write.barcode_files(my_sample_data_df, out_dir=tempdir())
}
